+++ assets/js/registry.js
@@ -24,6 +24,7 @@ import MenuController from './controllers/menu_controller.js';
 import ModalController from './controllers/modal_controller.js';
 import ModalTriggerController from './controllers/modal_trigger_controller.js';
 import PerPageController from './controllers/per_page_controller.js';
+import QuestionController from './controllers/question_controller.js';
 import ReadmoreController from './controllers/readmore_controller.js';
 import RevealController from './controllers/reveal_controller.js';
 import RevisionController from './controllers/revision_controller.js';
@@ -55,6 +56,7 @@ export const controllers = {
     'sonata-modal': ModalController,
     'sonata-modal-trigger': ModalTriggerController,
     'sonata-per-page': PerPageController,
+    'sonata-question': QuestionController,
     'sonata-readmore': ReadmoreController,
     'sonata-reveal': RevealController,
     'sonata-revision': RevisionController,